Some reconfiguring.

+++ src/com/Customer.java
@@ -9,7 +9,7 @@ public class Customer {
     private LinkedHashMap<String, String> details = new LinkedHashMap<>();
 
 
-    public Customer(String firstName, String lastName, String email, int ssn) {
+    public Customer(String firstName, String lastName, String email, long ssn) {
         details.put("firstname", firstName);
         details.put("lastname", lastName);
         details.put("email", email);
@@ -41,11 +41,11 @@ public class Customer {
         this.details.put("email", email);
     }
 
-    public int getSocialSecurityNumber (){
-        return Integer.parseInt(details.get("ssn"));
+    public long getSocialSecurityNumber (){
+        return Long.parseLong(details.get("ssn"));
     }
 
-    public void setSocialSecurityNumber (int socialSecurityNumber){
+    public void setSocialSecurityNumber (long socialSecurityNumber){
         this.details.put("ssn", String.valueOf(socialSecurityNumber));
     }
